refactor(configfile): split parsers into dedicated files
